Renaming (#3)
* Renaming

Renamed several DYNs/JPGs to reflect how they are actually named in the JSON file. This was an error carried over from previous naming convention. Also fixed some folder structures. This represents a check of the Dictionary through the input nodes.

* Rename through Math

Continued rename effort through Math category. This includes renames in the JSON because of invalid characters

* Rename through Script and String

This represents the rename effort through the rest of the nodes (excluding Revit nodes as there seems to be another issue at hand)

* Additions

Added color range and color palette that were missing.

* Renaming rest of Revit nodes

Finished renaming the Revit nodes to support what they actually are in the JSON.

* Update Dynamo_Nodes_Documentation.json